+++ src/test/java/de/devbliss/apitester/PosterIntegrationTest.java
@@ -1,16 +1,20 @@
 package de.devbliss.apitester;
 
 import static junit.framework.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 
+import java.io.IOException;
 import java.net.URI;
 
 import org.apache.http.HttpStatus;
+import org.apache.http.client.methods.HttpPost;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
 import de.devbliss.apitester.dummyserver.DummyApiServer;
 import de.devbliss.apitester.dummyserver.DummyDto;
+import de.devbliss.apitester.factory.PostFactory;
 import de.devbliss.apitester.factory.impl.DefaultPostFactory;
 
 /**
@@ -22,6 +26,10 @@ import de.devbliss.apitester.factory.impl.DefaultPostFactory;
  */
 public class PosterIntegrationTest {
 
+    private static final String HEADER_VALUE1 = "header_value1";
+    private static final String HEADER_NAME1 = "header_name1";
+    private static final String HEADER_VALUE2 = "header_value2";
+    private static final String HEADER_NAME2 = "header_name2";
     private DummyApiServer server;
 
     @Before
@@ -38,51 +46,69 @@ public class PosterIntegrationTest {
     @Test
     public void testPostOk() throws Exception {
         URI uri = server.buildGetRequestUri(HttpStatus.SC_OK);
-        ApiResponse response = Poster.post(uri);
+        Context wrapper = Poster.post(uri);
+        ApiRequest request = wrapper.apiRequest;
+        ApiResponse response = wrapper.apiResponse;
         ApiTestUtil.assertOk(response);
+        assertNotNull(request.headers);
     }
 
     @Test
     public void testPostOkWithPayload() throws Exception {
         DummyDto payload = createPayload();
         URI uri = server.buildGetRequestUri(HttpStatus.SC_OK);
-        ApiResponse response = Poster.post(uri, payload);
+        Context wrapper = Poster.post(uri, payload);
+        ApiRequest request = wrapper.apiRequest;
+        ApiResponse response = wrapper.apiResponse;
         ApiTestUtil.assertOk(response);
         DummyDto result = response.payloadJsonAs(DummyDto.class);
         assertEquals(payload, result);
+        assertNotNull(request.headers);
     }
 
     @Test
     public void testPostOkWithOwnPostFactory() throws Exception {
         URI uri = server.buildGetRequestUri(HttpStatus.SC_OK);
-        ApiResponse response = Poster.post(uri, new DefaultPostFactory());
+        Context wrapper = Poster.post(uri, new DefaultPostFactory());
+        ApiResponse response = wrapper.apiResponse;
+        ApiRequest request = wrapper.apiRequest;
         ApiTestUtil.assertOk(response);
+        assertNotNull(request.headers);
     }
 
     @Test
     public void testPostOkWithOwnTestState() throws Exception {
         URI uri = server.buildGetRequestUri(HttpStatus.SC_OK);
         TestState testState = ApiTesterModule.createTestState();
-        ApiResponse response = Poster.post(uri, testState);
+        Context wrapper = Poster.post(uri, testState);
+        ApiResponse response = wrapper.apiResponse;
+        ApiRequest request = wrapper.apiRequest;
         ApiTestUtil.assertOk(response);
+        assertNotNull(request.headers);
     }
 
     @Test
     public void testPostOkWithOwnPostFactoryAndTestState() throws Exception {
         URI uri = server.buildGetRequestUri(HttpStatus.SC_OK);
         TestState testState = ApiTesterModule.createTestState();
-        ApiResponse response = Poster.post(uri, new DefaultPostFactory(), testState);
+        Context wrapper = Poster.post(uri, new DefaultPostFactory(), testState);
+        ApiResponse response = wrapper.apiResponse;
+        ApiRequest request = wrapper.apiRequest;
         ApiTestUtil.assertOk(response);
+        assertNotNull(request.headers);
     }
 
     @Test
     public void testPostOkWithPayloadAndOwnPostFactory() throws Exception {
         DummyDto payload = createPayload();
         URI uri = server.buildGetRequestUri(HttpStatus.SC_OK);
-        ApiResponse response = Poster.post(uri, payload, new DefaultPostFactory());
+        Context wrapper = Poster.post(uri, payload, new DefaultPostFactory());
+        ApiResponse response = wrapper.apiResponse;
+        ApiRequest request = wrapper.apiRequest;
         ApiTestUtil.assertOk(response);
         DummyDto result = response.payloadJsonAs(DummyDto.class);
         assertEquals(payload, result);
+        assertNotNull(request.headers);
     }
 
     @Test
@@ -90,10 +116,13 @@ public class PosterIntegrationTest {
         DummyDto payload = createPayload();
         URI uri = server.buildGetRequestUri(HttpStatus.SC_OK);
         TestState testState = ApiTesterModule.createTestState();
-        ApiResponse response = Poster.post(uri, payload, testState);
+        Context wrapper = Poster.post(uri, payload, testState);
+        ApiResponse response = wrapper.apiResponse;
+        ApiRequest request = wrapper.apiRequest;
         ApiTestUtil.assertOk(response);
         DummyDto result = response.payloadJsonAs(DummyDto.class);
         assertEquals(payload, result);
+        assertNotNull(request.headers);
     }
 
     @Test
@@ -101,10 +130,39 @@ public class PosterIntegrationTest {
         DummyDto payload = createPayload();
         URI uri = server.buildGetRequestUri(HttpStatus.SC_OK);
         TestState testState = ApiTesterModule.createTestState();
-        ApiResponse response = Poster.post(uri, payload, testState, new DefaultPostFactory());
+        Context wrapper = Poster.post(uri, payload, testState, new DefaultPostFactory());
+        ApiResponse response = wrapper.apiResponse;
+        ApiRequest request = wrapper.apiRequest;
         ApiTestUtil.assertOk(response);
         DummyDto result = response.payloadJsonAs(DummyDto.class);
         assertEquals(payload, result);
+        assertNotNull(request.headers);
+    }
+
+    @Test
+    public void testPostWithHeaders() throws Exception {
+        DummyDto payload = createPayload();
+        URI uri = server.buildGetRequestUri(HttpStatus.SC_OK);
+        TestState testState = ApiTesterModule.createTestState();
+        Context wrapper = Poster.post(uri, payload, testState, getPostFactoryWithHeaders());
+        ApiResponse response = wrapper.apiResponse;
+        ApiRequest request = wrapper.apiRequest;
+        ApiTestUtil.assertOk(response);
+        assertEquals(2, request.headers.keySet().size());
+        assertEquals(HEADER_VALUE1, request.getHeader(HEADER_NAME1));
+        assertEquals(HEADER_VALUE2, request.getHeader(HEADER_NAME2));
+    }
+
+    private PostFactory getPostFactoryWithHeaders() {
+        return new PostFactory() {
+
+            public HttpPost createPostRequest(URI uri, Object payload) throws IOException {
+                HttpPost request = new HttpPost(uri);
+                request.setHeader(HEADER_NAME1, HEADER_VALUE1);
+                request.setHeader(HEADER_NAME2, HEADER_VALUE2);
+                return request;
+            }
+        };
     }
 
     private DummyDto createPayload() {